Base model refactor.
PiperOrigin-RevId: 306597558
